Make list entries (children) bounce relative to the box' (parent) movement
This is exciting! It's is the first real implementation that works.
Although there is still a lot to do such as 1) tracking the elements in
the viewport because currently the bounce is alwazs relative to the
first element in the list which when you scroll down is not ideal, 2)
decouple the acceleration from the deceleration so that if you stop
scrolling fast the children still bounce naturally and 3) add a
acceleration stack for smoother scroll movement