Add topViewPortIndex
Dynamically switch the element that the bounce is relative to. This
makes it possible to scroll down and still have it work.

Also added some text between th list items to see the effect more and
numbered the list items correclty.
The code base is currently pretty messy and needs to be cleaned up.